refactor(8766) separated the system client test to run independently
